+++ src/AP_Factory.cs
@@ -13,38 +13,52 @@ namespace TDriver {
     ///     DPA-99999-99999.doc                                                     (DPA Email, DPA Mail or DPA Fax Document)
     ///     DPA-99999-99999-For-NAME_HERE-FaxTo-1-999-999-9999-ATTN-NAME_HERE.doc   (DPA Fax Document)
     ///     DPA-99999-99999-For-NAME_HERE-FaxTo-1-999-999-9999                      (DPA Fax Document)
+    ///     99999-99999_faxto-1-999-999-9999.doc                                    (DPA2 Fax Document)
     /// </remarks>
     public static class AP_Factory {
         private static readonly Parser parser = new Parser();
 
-        internal static AP_Document Create(string file, AP_Subsection subsection) {
+        public static AP_Document Create(string file, AP_Subsection subsection) {
             //Determine the document type by checking against the general naming conventions
             DocumentType docType = DetermineDocumentType(file);
             if (docType == DocumentType.ERROR) {
-                Logger.AddError(Settings.ErrorLogfile, file + "was skipped. Does NOT follow general naming convention.");
+                Logger.AddError(Settings.ErrorLogfile, file + " was skipped. Does NOT follow general naming convention.");
                 return null;
             }
 
             //Open document with RTF Parser
-            //Todo load doc into parser and parse each item seperately
             AP_Document newApDoc = parser.FindData(file, docType);
             if (newApDoc == null) {
-                Logger.AddError(Settings.ErrorLogfile, file + "was skipped by the parser.");
+                Logger.AddError(Settings.ErrorLogfile, file + " was skipped by the parser. Verify the formatting.");
                 return null;
             }
 
-            //Derived case specific options
-            //DPA
-            //Add the type of DPA.
+            //Manage derived class, specific options
+            //**DPA***
             if (docType == DocumentType.DPA && newApDoc is DPA) {
+                //Add the type of DPA to the class.
                 ((DPA)newApDoc).KindOfDPA = subsection.Name;
             }
+            //**CME***
+            if (docType == DocumentType.CME && newApDoc is MedicalCME) {
+                //Verify the zip code to the specified region.
+                //TODO Add region in settings.
+                if (!((MedicalCME) newApDoc).CheckRegion("UPSTATE")) {
+                    //When the region doesn't match the specified.
+                    //Log it and mark as invalid.
+                    Logger.AddError(Settings.ErrorLogfile, file + " is NOT an " + "UPSTATE" + " zip code. File skipped."  );
+                    return null;
+                }
+
+            }
+
 
             //Determine the delivery method by analyzing the "SendTo" field's contents.
-            //TODO: Use fax number from file name?
             Tuple<DeliveryMethodType, string> result = DetermineDeliveryMethod(newApDoc.SendTo);
-            if (result == null)
+            if (result == null) {
+                Logger.AddError(Settings.ErrorLogfile, file + " was skipped. Couldn't determine who it's intended for.");
                 return null;
+            }
 
             //Change the delivery method and update where it's going to.
             newApDoc.ChangeDeliveryType(result.Item1);
@@ -57,15 +71,15 @@ namespace TDriver {
         ///     Find the delivery method for the document
         /// </summary>
         /// <param name="sendToField"></param>
-        /// <returns></returns>
+        /// <returns>Tuple(DeliveryMethod, Recipient)</returns>
         private static Tuple<DeliveryMethodType, string> DetermineDeliveryMethod(string sendToField) {
             //Check for email address.
             const String rgxEmailPattern =
                 @"[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?";
             var rgx = new Regex(rgxEmailPattern, RegexOptions.IgnoreCase);
-            MatchCollection emailMatches = rgx.Matches(sendToField);
-            if (emailMatches.Count == 1) {
-                return Tuple.Create(DeliveryMethodType.Email, emailMatches[0].Value);
+            Match emailMatch = rgx.Match(sendToField);
+            if (emailMatch.Success) {
+                return Tuple.Create(DeliveryMethodType.Email, emailMatch.Value);
             }
 
             //Check for a mailing address.
@@ -73,12 +87,12 @@ namespace TDriver {
             //ex. ' NY 13219'
             const String rgxAddressPattern = @"\s[a-z]{2}\s\d{5}";
             rgx = new Regex(rgxAddressPattern, RegexOptions.IgnoreCase);
-            MatchCollection mailMatches = rgx.Matches(sendToField);
-            if (mailMatches.Count == 1) {
+            Match mailMatch = rgx.Match(sendToField);
+            if (mailMatch.Success) {
                 return Tuple.Create(DeliveryMethodType.Mail, sendToField);
             }
 
-            //Check for fax number
+            //Check for VALID fax number
             // Return # found plus US-Code (1)
             const String rgxFaxPattern = @"\d{3}-\d{3}-\d{4}";
             rgx = new Regex(rgxFaxPattern, RegexOptions.IgnoreCase);
@@ -98,8 +112,8 @@ namespace TDriver {
         private static DocumentType DetermineDocumentType(string file) {
             //Check for Medical CME
             //Looks for 'Acct-99999-99999'
-            const String rgxCME_Convention = @"Acct-\d{5}-\d{5}";
-            var rgx = new Regex(rgxCME_Convention, RegexOptions.IgnoreCase);
+            const String rgxCMEConvention = @"Acct-\d{5}-\d{5}";
+            var rgx = new Regex(rgxCMEConvention, RegexOptions.IgnoreCase);
             Match cmeMatch = rgx.Match(file);
             if (cmeMatch.Success) {
                 return DocumentType.CME;
@@ -107,8 +121,8 @@ namespace TDriver {
 
             //Check for DPA
             //Looks for 'DPA-99999-99999-'
-            const String rgxDPA_Convention = @"DPA-\d{5}-\d{5}";
-            rgx = new Regex(rgxDPA_Convention, RegexOptions.IgnoreCase);
+            const String rgxDPAConvention = @"DPA-\d{5}-\d{5}";
+            rgx = new Regex(rgxDPAConvention, RegexOptions.IgnoreCase);
             Match dpaMatch = rgx.Match(file);
             if (dpaMatch.Success) {
                 return DocumentType.DPA;