+++ src/Logger.cs
@@ -4,27 +4,7 @@ using System.IO;
 
 namespace TDriver {
     internal static class Logger {
-        public static int ErrorCount = 0;
-
-        /// <summary>
-        ///     Add data to Comma Seperated Values file for data collecting.
-        /// </summary>
-        /// <param name="csvFile"></param>
-        /// <param name="subsection"></param>
-        /// <param name="deliveryMethod"></param>
-        /// <param name="accountNumber"></param>
-        public static void AddToCSV(String csvFile, String subsection, String deliveryMethod, String accountNumber) {
-            DateTime logTime = DateTime.Now;
-
-            if (!File.Exists(csvFile)) {
-                FileStream fs = File.Create(csvFile);
-                fs.Close();
-            }
-            string logAction = String.Format("{0},{1},{2},{3}{4}",
-                logTime.ToString(CultureInfo.InvariantCulture), subsection, deliveryMethod, accountNumber,
-                Environment.NewLine);
-            File.AppendAllText(csvFile, logAction);
-        }
+        public static int ErrorCount;
 
         /// <summary>
         ///     Logs an error message.