+++ src/MailWork.cs
@@ -12,14 +12,12 @@ namespace TDriver {
 
         public override AP_Document DocObject => _mail;
 
-        public string FileToPrint {
-            get { return _mail.FileToSend; }
-        }
+        public string FileToPrint => _mail.FileToSend;
 
         public override bool Process() {
 #if DEBUG //Allow simulating
-            //Debug result :: Mail Success.
-            //_mail.AddSentTime();
+    //Debug result :: Mail Success.
+            _mail.AddSentTime();
             
             Debug.WriteLine(String.Format("Mailed {0} to {1} for {2} with account {3} using printer: {4}", FileToPrint, _mail.SendTo, _mail.CustomerName, _mail.Account, "TODO Add Printer"));
             return true;