+++ src/Main.cs
@@ -9,7 +9,7 @@ using System.Windows.Forms;
 
 namespace TDriver {
     public partial class Main : Form {
-        private WorkQueue _dpaWorkQueue;
+        private WorkQueue _docWorkQueue;
         private List<Watcher> _folderWatchList;
 
         public Main() {
@@ -48,7 +48,7 @@ namespace TDriver {
         }
 
         /// <summary>
-        /// Fills and resizes the WatchFolderListview
+        ///     Fills and resizes the WatchFolderListview
         /// </summary>
         private void SetupWatchFolderListview() {
             Settings.WatchList.ForEach(delegate(AP_Subsection subSection) {
@@ -77,57 +77,66 @@ namespace TDriver {
         private Boolean _firstRun = true;
 
         private void tsBtnStart_Click(object sender, EventArgs e) {
-            //Start the watcher
             tbtnStart.Enabled = false;
+            
 
-            // Create the queue and watchlist when its the first time started.
+            // Create the queue and watcher list the first time started it'S.
             if (_firstRun) {
-                _dpaWorkQueue = new WorkQueue(Settings.DatabaseFile);
                 _folderWatchList = new List<Watcher>(Settings.MAX_WATCHLIST_SIZE);
+                _docWorkQueue = new WorkQueue(Settings.DatabaseFile);
+                
             }
 
             //Start the DPA Queue Worker
-            _dpaWorkQueue.StartQWorker();
+            _docWorkQueue.StartQWorker();
 
+            //Queue Directories and create the watchers
             foreach (AP_Subsection subsection in Settings.WatchList) {
-                //Queue Existing Files in the folder
-                _dpaWorkQueue.QueueDirectory(subsection.WatchFolder, subsection);
+                if (subsection.IsValid == false) continue;
+                if (Directory.Exists(subsection.WatchFolder)) {
+                    //Queue Existing Files in the folder
+                    _docWorkQueue.QueueDirectory(subsection.WatchFolder, subsection);
 
-                //Setup watcher for the folder.
-                if (_firstRun) {
-                    _folderWatchList.Add(new Watcher(subsection.WatchFolder, subsection, ref _dpaWorkQueue,
-                        Settings.FileDelayTime));
+                    //Setup watcher for the folder.
+                    _folderWatchList.Add(new Watcher(subsection.WatchFolder, subsection, ref _docWorkQueue, Settings.FileDelayTime));
+                    }
+            }
 
-                    //TODO Update UI with folders being watched.
+            //Start each folder watcher.
+            if (_folderWatchList.Any()) {
+                foreach (Watcher watcher in _folderWatchList) {
+                    watcher.Start();
                 }
+
+                //Update local variables
+                _firstRun = false;
+
+                //Update UI
+                tbtnStop.Enabled = true;
+                tslblStatus.Text = "Running";
+                tslblStatus.ForeColor = Color.Green;
             }
-
-            //Start each foloder watcher.
-            foreach (Watcher watcher in _folderWatchList) {
-                watcher.Start();
+            else {
+                _docWorkQueue.StopQWorker();
+                MessageBox.Show("No valid folder watchers. Please check settings and try again.");
+                tbtnStart.Enabled = true;
             }
-
-            //Update local variables
-            _firstRun = false;
-
-            //Update UI
-            tbtnStop.Enabled = true;
-            tslblStatus.Text = "Running";
-            tslblStatus.ForeColor = Color.Green;
         }
 
         private void tbtnStop_Click(object sender, EventArgs e) {
             tbtnStop.Enabled = false;
 
-            //Stop each folder watcher
+            //Stop each folder watcher and clear list.
             foreach (Watcher watcher in _folderWatchList) {
                 watcher.Stop();
             }
+            _folderWatchList.Clear();
+
             //Stop DPA Queue Worker
-            _dpaWorkQueue.StopQWorker();
+            _docWorkQueue.StopQWorker();
 
             //Update UI
-            Debug.WriteLine("Poll haulted.");
+            Debug.WriteLine("Watchers halted.");
             tslblStatus.Text = "Stopped";
             tslblStatus.ForeColor = Color.DarkRed;
             tbtnStart.Enabled = true;
@@ -135,10 +144,10 @@ namespace TDriver {
 
 
         private void frmMain_Resize(object sender, EventArgs e) {
-            // Minimize to traybar. (Taskbar is very crowded at work)
+            // Minimize to tray bar. (Task-bar is very crowded at work)
             if (FormWindowState.Minimized == WindowState) {
                 notifyIcon1.BalloonTipTitle = "T: Driver";
-                notifyIcon1.BalloonTipText = "Minimized to traybar.";
+                notifyIcon1.BalloonTipText = "Minimized to tray bar.";
                 //notifyIcon1.ShowBalloonTip(50);
                 //ShowInTaskbar = false;
             }
@@ -166,5 +175,22 @@ namespace TDriver {
         }
 
         #endregion
+
+        private void MainFormClosing(object sender, FormClosingEventArgs e) {
+            if (_folderWatchList  != null) {
+                //Stop each folder watcher and clear list.
+                foreach (Watcher watcher in _folderWatchList) {
+                    watcher.Stop();
+                }
+                _folderWatchList.Clear();
+            }
+
+            //Dispose work queue and queue worker.
+            if (_docWorkQueue != null) {
+                if (_docWorkQueue.IsRunning)
+                    _docWorkQueue.StopQWorker();
+                _docWorkQueue.Dispose();
+            }
+        }
     }
 }