+++ src/Watcher.cs
@@ -2,11 +2,12 @@
 using System.Diagnostics;
 using System.IO;
 using System.Timers;
+using FileSystemWatcherEx;
 
 namespace TDriver {
     public class Watcher {
         private readonly int _fileDelay;
-        private readonly FileSystemWatcher _watcher;
+        private readonly WatcherEx _watcher;
         private readonly WorkQueue _workQueue;
 
         /// <summary>
@@ -17,20 +18,33 @@ namespace TDriver {
         /// <param name="workQueue">Queue for worked to be added to.</param>
         /// <param name="delay">Delay time in milliseconds, </param>
         public Watcher(string sPath, AP_Subsection folderSubsection, ref WorkQueue workQueue, int delay) {
+
+            _fileDelay = delay;
+            _workQueue = workQueue;
+
+            //Check if the directory exists.
+            if (!Directory.Exists(sPath)) {
+                Logger.AddError(Settings.ErrorLogfile, sPath + " does not exist!");
+                return;
+            }
             try {
-                _fileDelay = delay;
-                _workQueue = workQueue;
-                //Check if the directory exists.
-                if (!Directory.Exists(sPath)) {
-                    //Form.LogError(sPath + " does not exist!");
-                    return;
-                }
+                
+                // Setup the watcher info.
+                WatcherInfo wInfo = new WatcherInfo {
+                    ChangesFilters = NotifyFilters.Size,
+                    IncludeSubFolders = false,
+                    WatchesFilters = WatcherChangeTypes.Created,
+                    WatchForDisposed = true,
+                    WatchForError = false,
+                    WatchPath = sPath, //Path to Watch
+                    BufferKBytes = 8, //Default Buffer
+                    MonitorPathInterval = 2*60*1000, //Check folder availability every two minutes.
+                    FileFilter = "*.doc"
+                };
 
                 // Watch the directory for new word documents.
-                _watcher = new FileSystemWatcher(sPath, "*.doc") {
-                    NotifyFilter = NotifyFilters.FileName
-                };
-                _watcher.Created += (sender, e) => NewFileCreated(e.FullPath, folderSubsection);
+                _watcher = new WatcherEx(wInfo);
+                _watcher.EventCreated += (sender, e) => { NewFileCreated(((FileSystemEventArgs) (e.Arguments)).FullPath, folderSubsection); };
             }
             catch (Exception ex) {
                 Debug.WriteLine(ex.Message);
@@ -50,25 +64,24 @@ namespace TDriver {
             }
 
             Debug.WriteLine("New File detected!");
-            //An artifical wait to handle duplicate file creations bug from our web application.
+            // An artificial wait to handle duplicate file creations bug from our web application.
             // Multiple FileDelay by 1000 to get milliseconds.
-            //BUG check for resource usage for high volume of undisposed timers.
             var aTimer = new Timer((double) _fileDelay*1000) {AutoReset = false};
-
             aTimer.Elapsed += (sender, e) => _workQueue.FoundFileCheck(file, fileSubsection);
             aTimer.Enabled = true;
         }
 
+        ///Stops the watcher (and disposes)
         public void Stop() {
-            _watcher.EnableRaisingEvents = false;
+            _watcher.Stop();
+            //_watcher.Dispose();
         }
 
+        /// <summary>
+        /// Starts the Watcher
+        /// </summary>
         public void Start() {
-            _watcher.EnableRaisingEvents = true;
-        }
-
-        public void Dispose() {
-            _watcher.Dispose();
+            _watcher.Start();
         }
     }
 }