+++ src/lib/FileSystemWatcherEx/WatcherCommon.cs
@@ -1,0 +1,12 @@
+﻿// http://www.codeproject.com/Articles/58740/FileSystemWatcher-Pure-Chaos-Part-of
+namespace FileSystemWatcherEx
+{
+	////////////////////////////////////////////////////////////////////////////////////// 
+	////////////////////////////////////////////////////////////////////////////////////// 
+	/// <summary>
+	/// This enum is used to indicate which argument type is valid in the WatcherEventArgs 
+	/// object.
+	/// </summary>
+	public enum ArgumentType { FileSystem, Renamed, Error, StandardEvent, PathAvailability };
+
+}